fix: bug fixes and cleanups related to AppConfigurator
